+++ odoo/local-src/sf_rma/models/rma.py
@@ -308,12 +308,17 @@ class RMA(models.Model):
             limit=1
         )
 
+        # Operation type for RMA reception
+        comp = self.env.user.company_id
+        if not comp.rma_receive_picking_type_id:
+            raise UserError(_('Configure RMA reception operation type.'))
+
         return {
             'partner_id': self.partner_id.id,
             'rma_id': self.id,
             'origin': self.name,
             'company_id': self.company_id.id,
-            'picking_type_id': self.env.ref('sf_rma.picking_type_rma').id,
+            'picking_type_id': comp.rma_receive_picking_type_id.id,
             'location_dest_id': warehouse.lot_rma_id.id,
         }