+++ odoo/local-src/sf_rma/tests/test_rma.py
@@ -23,6 +23,10 @@ class TestRMA(TransactionCase):
             'decision': 'free',
         })
 
+        comp = self.env.user.company_id
+        comp.rma_receive_picking_type_id = \
+            self.env.ref('sf_rma.picking_type_rma')
+
     def test_rma_sequence(self):
         self.assertEquals(self.rma.name[:3], 'RMA')