+++ HISTORY.rst
@@ -32,6 +32,7 @@ latest (unreleased)
 * BSSFL-443: RMA number is readonly
 * BSSFL-491: Fix invoice report (add condition to display bank details)
 * BSSFL-489: Don't duplicate confirmation date on purchase order
+* BSSFL-483: RMA Inc reception
 
 **Build**